+++ src/app/dashboard/campaigns/[id]/layout.tsx
@@ -1,11 +1,22 @@
-import { generateStaticParams } from './generateStaticParams'
+import { supabase } from '@/lib/supabase/client'
 
-export { generateStaticParams }
+// Esta função é necessária para exportação estática com Next.js
+export async function generateStaticParams() {
+  // Buscar todos os IDs de campanhas do Supabase
+  const { data } = await supabase
+    .from('campaigns')
+    .select('id')
+  
+  // Retornar um array de objetos com o parâmetro id
+  return (data || []).map((campaign: { id: any }) => ({
+    id: campaign.id.toString(),
+  }))
+}
 
 export default function CampaignLayout({
   children,
 }: {
-  children: React.ReactNode
+  children: any
 }) {
   return <>{children}</>
 }
